+++ Source/SelfService/Web/App.tsx
@@ -7,8 +7,8 @@ import { IBreadcrumbItem } from '@fluentui/react/lib/Breadcrumb';
 import { CommandBar, ICommandBarItemProps } from '@fluentui/react/lib/CommandBar';
 import { Route, BrowserRouter, useHistory } from 'react-router-dom';
 import { WebhooksConfig } from './micoservice/businessMomentsAdaptor/configuration/WebhooksConfiguration';
-import { Editor as BusinessMomentEditor } from './moments/BusinessMomentEditor';
-import { Editor as EntityEditor } from './moments/EntityEditor';
+import { Editor as BusinessMomentEditor } from './businessMoments/Editor';
+import { Editor as EntityEditor } from './businessMoments/EntityEditor';
 import { ApplicationScreen } from './screens/applicationScreen';
 import { ApplicationsScreen } from './screens/applicationsScreen';
 import { QuickScreen } from './screens/quickScreen';
